error: Split into multiple modules.
```
git diff HEAD^1:src/error.rs src/error/key_rejected.rs
git diff HEAD^1:src/error.rs src/error/into_unspecified.rs
git diff HEAD^1:src/error.rs src/error/mod.rs
git diff HEAD^1:src/error.rs src/error/unspecified.rs
```
